move messageCapable to dedicated file
